Cleanup import cycle
